chat-backend/requirements.txt
numpy
requests
pymupdf4llm[ocr,layout]
sentence_transformers
faiss-cpu
fastapi[standard]
pydantic
openpyxl
python-docx
